Struct firebase::Firebase

pub struct Firebase { /* fields omitted */ }

A Firebase instance to manage data.

Methods

impl Firebase

fn new(url: &str) -> Result<Self, ParseError>

Creates a new Firebase instance from the url of the fb server The url should be a valid HTTPS url, anything else will result in an error:

https://docs-examples.firebaseio.com/

Failures

If a url is not specified with the HTTPS scheme, a Err(ParseError::UrlIsNotHTTPS) will be returned.

If a url cannot be parsed into a valid url then a Err(ParseError::Parser(url::ParseError) will be returned.

fn from_url(url: &Url) -> Result<Self, ParseError>

Creates a firebase reference from a borrow of a Url instance.

Failures

If a url is not specified with the HTTPS scheme, a Err(ParseError::UrlIsNotHTTPS) will be returned.

If a url cannot be parsed into a valid url then a Err(ParseError::Parser(url::ParseError) will be returned.

fn at(&self, add_path: &str) -> Result<Self, ParseError>

Creates a new firebase instance that extends the path of an old firebase instance. Each time a reference is created a clone of the Firebase instance if done, all Firebase instances follow this immutable style.

#Examples

// Points to the root of the db ( / )
let fb = Firebase::new("https://myfb.firebaseio.com").unwrap();
// A new reference to /friends/yasha
let yasha = fb.at("/friends/yasha").unwrap();
// A new reference to /friends/yasha/messages
let messages = yasha.at("messages").unwrap();

fn ops(&self, opts: &FbOps) -> FirebaseParams

Creates a FirebaseParams instance, this instance has query parameters that are associated with it and that are used in every request made. Since query parameters only affect incomming data from Firebase, you can only GET data with a FirebaseParams instance.

This constructor takes in a FbOps struct to define all of its parameters, all undefined parameters can be omitted by extending the new FbOps struct by its default.

Examples

let fb = Firebase::new("https://db.fb.com").unwrap();
let query = fb.ops(&FbOps {
   order_by:       Some("Hello World"),
   limit_to_first: Some(5),
   end_at:         Some(7),
   equal_to:       Some(3),
   shallow:        Some(true),
   format:         Some(true),
   .. FbOps::default()
});

fn get_url(&self) -> &str

Returns the current URL as a string that will be used to make the REST call when talking to Firebase.

fn get(&self) -> Result<Response, ReqErr>

Gets data from Firebase.

Examples

let firebase = Firebase::new("https://shows.firebaseio.com").unwrap();
let episode = firebase.at("/futurama/episodes/140").unwrap();
let info = episode.get();

fn set(&self, data: &str) -> Result<Response, ReqErr>

Sets data to Firebase.

Examples

let firebase = Firebase::new("https://shows.firebaseio.com").unwrap();
let episode = firebase.at("/futurama/episodes/140/description").unwrap();
let info = episode.set("The Last Episode!");

fn push(&self, data: &str) -> Result<Response, ReqErr>

Pushes data to Firebase.

Examples

let firebase = Firebase::new("https://shows.firebaseio.com").unwrap();
let episodes = firebase.at("/futurama/episodes").unwrap();
let info = episodes.push("The Lost Episode");

fn update(&self, data: &str) -> Result<Response, ReqErr>

Updates Firebase data.

Examples

let firebase = Firebase::new("https://shows.firebaseio.com").unwrap();
let desc = firebase.at("/futurama/episodes/140/description").unwrap();
let info = desc.update("The Penultimate Episode!");

fn remove(&self) -> Result<Response, ReqErr>

Removes Firebase data.

Examples

let firebase = Firebase::new("https://shows.firebaseio.com").unwrap();
let episode = firebase.at("/futurama/episodes/141").unwrap();
episode.remove();

fn get_async<F>(&self, callback: F) -> JoinHandle<()> where F: Fn(Result<Response, ReqErr>) + Send + 'static

Asynchronous version of the get method, takes a callback and returns a handle to the thread making the request to Firebase.

Examples

let firebase = Firebase::new("https://shows.firebaseio.com").unwrap();
let desc = firebase.at("/futurama/episodes/141/description").unwrap();
let original = "The Lost Episode";
desc.get_async(move |result| {
    if result.unwrap().body != original {
        println!("The description changed!");
    }
});

fn set_async<S, F>(&self, data: S, callback: F) -> JoinHandle<()> where F: Fn(Result<Response, ReqErr>) + Send + 'static, S: Into<String>

Asynchronous version of the set method, takes a callback and returns a handle to the thread making the request to Firebase.

fn push_async<S, F>(&self, data: S, callback: F) -> JoinHandle<()> where F: Fn(Result<Response, ReqErr>) + Send + 'static, S: Into<String>

Asynchronous version of the push method, takes a callback and returns a handle to the thread making the request to Firebase.

fn update_async<S, F>(&self, data: S, callback: F) -> JoinHandle<()> where F: Fn(Result<Response, ReqErr>) + Send + 'static, S: Into<String>

Asynchronous version of the update method, takes a callback and returns a handle to the thread making the request to Firebase.

fn remove_async<F>(&self, callback: F) -> JoinHandle<()> where F: Fn(Result<Response, ReqErr>) + Send + 'static

Asynchronous version of the remove method, takes a callback and returns a handle to the thread making the request to Firebase.

fn order_by(&self, key: &str) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and sorts this data by the key provided.

fn limit_to_first(&self, count: u32) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and limits the number of entries returned on each request to the first count. Often used with order_by.

fn limit_to_last(&self, count: u32) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and limits the number of entries returned on each request to the last count. Often used with order_by.

fn start_at(&self, index: u32) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and only returns entries starting after the specified index. Often used with order_by.

fn end_at(&self, index: u32) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and only returns entries appearing before the specified index. Often used with order_by.

fn equal_to(&self, index: u32) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and returns only the entry at the specified index. Often used with order_by.

fn shallow(&self, flag: bool) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and only returns a shallow copy of the db in every request.

fn format(&self) -> FirebaseParams

Creates a FirebaseParams instance, a Firebase struct that only knows how to GET data, and formats the data to be exported in every request. (e.g. includes a priority field).